Move dedicated components in a dedicated directory
